firebase/functions/main.py
"""
Firebase Functions for contract processing pipeline.
"""
import json
import logging
from typing import Dict, Any
from firebase_functions import https_fn, storage_fn
from firebase_admin import initialize_app, firestore, storage
from google.cloud import aiplatform

from pipeline.orchestrator import ContractPipeline

# Initialize Firebase
initialize_app()
db = firestore.client()
bucket = storage.bucket()

# Initialize Vertex AI
aiplatform.init(project=settings.VERTEX_AI_PROJECT, location=settings.VERTEX_AI_LOCATION)

logger = logging.getLogger(__name__)


@storage_fn.on_object_finalized()
def process_uploaded_contract(cloud_event: storage_fn.CloudEvent) -> None:
    """
    Firebase Storage trigger - processes contracts when uploaded.
    
    Triggered when a new file is uploaded to Firebase Storage.
    """
    # TODO: Get file information from cloud_event
    file_name = cloud_event.data["name"]
    bucket_name = cloud_event.data["bucket"]
    
    # TODO: Validate file type (PDF only)
    if not file_name.lower().endswith('.pdf'):
        logger.info(f"Skipping non-PDF file: {file_name}")
        return
    
    # TODO: Download file from Storage
    blob = bucket.blob(file_name)
    local_path = f"/tmp/{file_name}"
    blob.download_to_filename(local_path)
    
    # TODO: Initialize processing pipeline
    pipeline = ContractPipeline()
    
    # TODO: Process contract
    result = pipeline.process_contract(local_path)
    
    # TODO: Store results in Firestore
    doc_id = file_name.replace('.pdf', '')
    db.collection('contracts').document(doc_id).set({
        'filename': file_name,
        'status': 'completed' if result['success'] else 'failed',
        'processed_at': firestore.SERVER_TIMESTAMP,
        'metadata': result.get('contract', {}).get('metadata', {}),
        'analysis': result.get('analysis', {}),
        'error': result.get('error')
    })
    
    logger.info(f"Contract processing completed: {file_name}")


@https_fn.on_request()
def process_contract_api(req: https_fn.Request) -> https_fn.Response:
    """
    HTTP endpoint for manual contract processing.
    
    POST /process-contract
    Body: { "file_path": "path/to/contract.pdf" }
    """
    # TODO: Validate request
    if req.method != 'POST':
        return https_fn.Response("Method not allowed", status=405)
    
    try:
        data = req.get_json()
        file_path = data.get('file_path')
        
        if not file_path:
            return https_fn.Response("Missing file_path", status=400)
        
        # TODO: Process contract
        pipeline = ContractPipeline()
        result = pipeline.process_contract(file_path)
        
        return https_fn.Response(
            json.dumps(result, default=str),
            status=200,
            headers={'Content-Type': 'application/json'}
        )
        
    except Exception as e:
        logger.error(f"API error: {str(e)}")
        return https_fn.Response(f"Error: {str(e)}", status=500)


@https_fn.on_request()
def query_contract_api(req: https_fn.Request) -> https_fn.Response:
    """
    HTTP endpoint for querying contracts using RAG.
    
    POST /query-contract
    Body: { "contract_id": "doc_id", "question": "What are the payment terms?" }
    """
    # TODO: Validate request
    if req.method != 'POST':
        return https_fn.Response("Method not allowed", status=405)
    
    try:
        data = req.get_json()
        contract_id = data.get('contract_id')
        question = data.get('question')
        
        if not contract_id or not question:
            return https_fn.Response("Missing contract_id or question", status=400)
        
        # TODO: Retrieve contract from Firestore
        doc_ref = db.collection('contracts').document(contract_id)
        doc = doc_ref.get()
        
        if not doc.exists:
            return https_fn.Response("Contract not found", status=404)
        
        # TODO: Use RAG to answer question
        pipeline = ContractPipeline()
        # Note: This would need the full contract object, not just metadata
        answer = "TODO: Implement RAG query with retrieved contract data"
        
        return https_fn.Response(
            json.dumps({'answer': answer}),
            status=200,
            headers={'Content-Type': 'application/json'}
        )
        
    except Exception as e:
        logger.error(f"Query error: {str(e)}")
        return https_fn.Response(f"Error: {str(e)}", status=500)


@https_fn.on_request()
def batch_process_api(req: https_fn.Request) -> https_fn.Response:
    """
    HTTP endpoint for batch processing multiple contracts.
    
    POST /batch-process
    Body: { "folder_path": "contracts/" }
    """
    # TODO: Implement batch processing endpoint
    # TODO: List files in Storage folder
    # TODO: Queue processing jobs
    # TODO: Return batch job ID for tracking
    pass