pipeline/ocr_extractor.py
"""
OCR and text extraction module for contract processing.
"""
import os
import logging
from typing import Tuple, Optional
from pathlib import Path
import fitz  # PyMuPDF
import pytesseract
import pdfplumber
from PIL import Image
from models.contract import ContractMetadata


class OCRExtractor:
    """Handles OCR and text extraction from PDF documents."""
    
    def __init__(self, tesseract_path: Optional[str] = None):
        """Initialize OCR extractor with optional Tesseract path."""
        if tesseract_path:
            pytesseract.pytesseract.tesseract_cmd = tesseract_path
        
        self.logger = logging.getLogger(__name__)
    
    def extract_text(self, file_path: str) -> Tuple[str, ContractMetadata]:
        """
        Main extraction method - determines best approach for PDF.
        
        Args:
            file_path: Path to PDF file
            
        Returns:
            Tuple of (extracted_text, metadata)
        """
        # TODO: Implement file type detection
        # TODO: Choose between OCR and text-based extraction
        # TODO: Handle different file formats (PDF, images, etc.)
        
        if self._is_text_based_pdf(file_path):
            return self._extract_with_pdfplumber(file_path)
        else:
            return self._extract_with_ocr(file_path)
    
    def _is_text_based_pdf(self, file_path: str) -> bool:
        """Check if PDF contains extractable text or needs OCR."""
        # TODO: Implement text detection logic
        # TODO: Check for embedded text vs scanned images
        pass
    
    def _extract_with_pdfplumber(self, file_path: str) -> Tuple[str, ContractMetadata]:
        """Extract text from text-based PDFs using pdfplumber."""
        # TODO: Open PDF with pdfplumber
        # TODO: Extract text while preserving layout
        # TODO: Handle tables and structured content
        # TODO: Generate metadata
        pass
    
    def _extract_with_ocr(self, file_path: str) -> Tuple[str, ContractMetadata]:
        """Extract text from scanned PDFs using Tesseract OCR."""
        # TODO: Convert PDF pages to images
        # TODO: Apply image preprocessing (deskew, denoise, etc.)
        # TODO: Run Tesseract OCR on each page
        # TODO: Combine results and calculate confidence scores
        pass
    
    def _preprocess_image(self, image: Image.Image) -> Image.Image:
        """Apply preprocessing to improve OCR accuracy."""
        # TODO: Implement image preprocessing pipeline:
        # - Deskewing
        # - Noise reduction
        # - Contrast enhancement
        # - Binarization
        pass
    
    def _extract_metadata(self, file_path: str) -> ContractMetadata:
        """Extract metadata from PDF file."""
        # TODO: Get file size, page count, creation date
        # TODO: Detect language
        # TODO: Calculate processing statistics
        pass